Bug 3850 - SFC Sonar: Breaking package dependencies
Introduced a new API class so that the Locking methods
can be outside the main OpendaylightSfc class. This breaks
one of the many package cycles

Changed all other files accordingly

+++ sfc-provider/src/main/java/org/opendaylight/sfc/provider/SfcProviderSfpEntryDataListener.java
@@ -11,6 +11,7 @@ package org.opendaylight.sfc.provider;
 
 import org.opendaylight.controller.md.sal.binding.api.DataChangeListener;
 import org.opendaylight.controller.md.sal.common.api.data.AsyncDataChangeEvent;
+import org.opendaylight.sfc.provider.api.SfcConcurrencyAPI;
 import org.opendaylight.sfc.provider.api.SfcProviderServicePathAPI;
 import org.opendaylight.yang.gen.v1.urn.cisco.params.xml.ns.yang.sfc.sfp.rev140701.service.function.paths.ServiceFunctionPath;
 import org.opendaylight.yangtools.yang.binding.DataObject;
@@ -24,16 +25,14 @@ import java.util.Set;
 import static org.opendaylight.sfc.provider.SfcProviderDebug.printTraceStart;
 import static org.opendaylight.sfc.provider.SfcProviderDebug.printTraceStop;
 
-//import org.opendaylight.yang.gen.v1.urn.cisco.params.xml.ns.yang.sfc.rsp.rev140701.rendered.service.paths.RenderedServicePath;
-//import javax.ws.rs.HttpMethod;
-
 /**
  * This class is the DataListener for SFP Entry changes.
+ * <p/>
+ * <p/>
  *
- * <p>
  * @author Reinaldo Penno (rapenno@gmail.com)
  * @version 0.1
- * @since       2014-06-30
+ * @since 2014-06-30
  */
 
 public class SfcProviderSfpEntryDataListener implements DataChangeListener {
@@ -43,32 +42,38 @@ public class SfcProviderSfpEntryDataListener implements DataChangeListener {
 
     @Override
     public void onDataChanged(
-            final AsyncDataChangeEvent<InstanceIdentifier<?>, DataObject> change ) {
+            final AsyncDataChangeEvent<InstanceIdentifier<?>, DataObject> change) {
 
         printTraceStart(LOG);
-        ODL_SFC.getLock();
+        if (SfcConcurrencyAPI.getLock()) {
+            try {
 
-        Map<InstanceIdentifier<?>, DataObject> dataOriginalDataObject = change.getOriginalData();
+                Map<InstanceIdentifier<?>, DataObject> dataOriginalDataObject = change.getOriginalData();
 
-        for (Map.Entry<InstanceIdentifier<?>, DataObject> entry : dataOriginalDataObject.entrySet()) {
-            if( entry.getValue() instanceof  ServiceFunctionPath) {
-                ServiceFunctionPath originalServiceFunctionPath = (ServiceFunctionPath) entry.getValue();
-                LOG.debug("\n########## Original Service path: {}",
-                        originalServiceFunctionPath.getName());
+                for (Map.Entry<InstanceIdentifier<?>, DataObject> entry : dataOriginalDataObject.entrySet()) {
+                    if (entry.getValue() instanceof ServiceFunctionPath) {
+                        ServiceFunctionPath originalServiceFunctionPath = (ServiceFunctionPath) entry.getValue();
+                        LOG.debug("\n########## Original Service path: {}",
+                                originalServiceFunctionPath.getName());
+                    }
+                }
+
+                // SFP DELETION
+                Set<InstanceIdentifier<?>> dataRemovedConfigurationIID = change.getRemovedPaths();
+                for (InstanceIdentifier instanceIdentifier : dataRemovedConfigurationIID) {
+                    DataObject dataObject = dataOriginalDataObject.get(instanceIdentifier);
+                    if (dataObject instanceof ServiceFunctionPath) {
+
+                        ServiceFunctionPath originalServiceFunctionPath = (ServiceFunctionPath) dataObject;
+                        SfcProviderServicePathAPI.deleteServicePathStateExecutor(originalServiceFunctionPath.getName());
+                    }
+                }
+            } finally {
+                SfcConcurrencyAPI.releaseLock();
             }
+        } else {
+            LOG.error("{}: Failed to Acquire Lock", Thread.currentThread().getStackTrace()[1]);
         }
-
-        // SFP DELETION
-        Set<InstanceIdentifier<?>> dataRemovedConfigurationIID = change.getRemovedPaths();
-        for (InstanceIdentifier instanceIdentifier : dataRemovedConfigurationIID) {
-            DataObject dataObject = dataOriginalDataObject.get(instanceIdentifier);
-            if( dataObject instanceof ServiceFunctionPath) {
-
-                ServiceFunctionPath originalServiceFunctionPath = (ServiceFunctionPath) dataObject;
-                SfcProviderServicePathAPI.deleteServicePathStateExecutor(originalServiceFunctionPath.getName());
-            }
-        }
-        ODL_SFC.releaseLock();
         printTraceStop(LOG);
     }
